Oprava: ArCommand volá správnou metodu AR() místo AD() + ArCommand kontroluje zůstatek před smazáním účtu

+++ src/Application/Commands/ArCommand.cs
@@ -5,15 +5,19 @@ using P2PBank.Utils;
 using P2PBank.Application;
 using P2PBank.Data.Interface;
 using P2PBank.Application.Interface.Models;
+using P2PBank.Presentation.Interface;
+using P2PBank.Presentation.Tcp;
 
 public class ArCommand : Command
 {
     private IAccountRepository accountRepository;
+    private ServerConfig serverConfig;
 
-    public ArCommand(IAccountRepository accountRepository, Log log) : base(log)
+    public ArCommand(IAccountRepository accountRepository, ServerConfig serverConfig, Log log) : base(log)
     {
         Name = "AR";
         this.accountRepository = accountRepository;
+        this.serverConfig = serverConfig;
     }
 
     public override void InternalExecute()
@@ -39,6 +43,11 @@ public class ArCommand : Command
                 throw new UnifiedMessageException("Account was not found.");
             }
 
+            if(acc.Balance != 0)
+            {
+                throw new UnifiedMessageException("Nelze smazat bankovní účet na kterém jsou finance.");
+            }
+
             try
             {
                 accountRepository.Delete(acc);
@@ -50,11 +59,14 @@ public class ArCommand : Command
         }
         else
         {
-            BankConnection con = new(ip);
+            TcpServerConfig? tcpConfig = serverConfig as TcpServerConfig;
+            int port = tcpConfig?.Port ?? 65525;
+            int timeout = tcpConfig?.ClientTimeout ?? 5;
+            BankConnection con = new(ip, port, timeout);
 
             try
             {
-                con.AD(id);
+                con.AR(id);
             }
             catch
             {
